+++ apps/customize_menu.sh
@@ -7,12 +7,11 @@ customize_menu() {
         set_title "macrift > customize"
         local choice
         choice=$(show_menu "Customize" \
-            "Terminal" \
+            "Terminal Emulator" \
             "Shell" \
             "FastFetch" \
             "Code Editor" \
             "Dock Layout" \
-            "Spicetify" \
             "---" \
             "Wallpaper" \
             "Back")
@@ -23,8 +22,7 @@ customize_menu() {
             3) source "$MACRIFT_DIR/apps/terminal.sh" && fastfetch_menu ;;
             4) source "$MACRIFT_DIR/apps/editor.sh" && editor_menu ;;
             5) source "$MACRIFT_DIR/apps/dock_layout.sh" && dock_layout_menu ;;
-            6) source "$MACRIFT_DIR/apps/spicetify.sh" && restore_marketplace ;;
-            7) source "$MACRIFT_DIR/apps/wallpaper.sh" && wallpaper_menu ;;
+            6) source "$MACRIFT_DIR/apps/wallpaper.sh" && wallpaper_menu ;;
             0) return ;;
             *) ;;
         esac